requirements.txt
appdirs==1.4.4
ariadne==0.12.0
astroid==2.4.2
backcall==0.2.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
graphql-core==3.0.5
h11==0.10.0
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
regex==2020.9.27
six==1.15.0
starlette==0.13.8
toml==0.10.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uvicorn==0.12.1
wcwidth==0.2.5
wrapt==1.12.1
